lib: use time validator function copied from PF4 in shutdown dialog
Our time validator would accept characters.

An issue to enhance the TimePicker component was opened, asking for
providing the validated status through the onChange callback.

See https://github.com/patternfly/patternfly-react/issues/5564
